libaitfoaq: Move active player into GamePhase
I avoided this in the beginning because it seemd more complex to write
the guard statements because comparisons/matches with enums holding data
is complexer, but this is the smaller evil. Holding an order of
buzzed-in contestants is way more complicated, and we can't rely on the
type system to ensure there's always at least one entry in that list
when there's supposed to be one. Yay Rust type safety!