rpi-controller: Add async handling and channels
The websocket communication should happen in a async runtime. It's
possible to do it in a blocking fashion, but I really don't wanna learn
new tricks now.

I thought about using interrupts with the GPIO at first, but then
figuring out which handset the interrupt belonged to was to much of a
hassle for me. So I just turned the control loop async with a tokio
interval, and that's it.