rpi-controller: Send appropriate connection msgs
If there is no player with our name, register, if there is, reconnect.
Do that in a loop until the server thinks this socket is controlling a
player.

We sleep for 1s in there to give the server some time to update it's
state. It's only on initial connection, it's fine.

We're just gonna raw-dog the json!() macro to generate our events here.
I'm out of time to care...